requirements.txt
streamlit
langchain-core
sentence-transformers
langchain-community 
faiss-cpu
langchain-google-genai
PyMuPDF
langchain-groq
langgraph
pypdf